Did some small, yet significant performance improvements, explained below:
- Using a Buffer.concat and generally reducing the amount of operations made by the baseId function.
- Used a bitwise operator for the generate -> count variable, as they are generally faster than arithmetic ones.
- Removed the pad function as its only usage can be converted, as did, to a sliced string.
- As the '==' padding used in the baseId function is always the same, I brought it out of the function scope preventing a redundant operation being done every time.


All tests seem to pass and benchmarks provide a +1/+5% benefit.

Before updates, on a Macbook Pro 2021 (M1 Max, 32GB RAM):
```
hyperid - variable length x 14,254,472 ops/sec ±0.62% (88 runs sampled)
hyperid - fixed length x 14,098,171 ops/sec ±0.69% (90 runs sampled)
hyperid - fixed length, url safe x 14,382,843 ops/sec ±0.75% (91 runs sampled)
```

After updates:
```
hyperid - variable length x 14,481,182 ops/sec ±0.52% (84 runs sampled)
hyperid - fixed length x 14,268,481 ops/sec ±0.57% (87 runs sampled)
hyperid - fixed length, url safe x 14,797,637 ops/sec ±0.62% (92 runs sampled)
```

I reached 15M ops/sec in a couple of benchmarks, yet those results are the medians.